Moved test helpers to their own directory
